Created loan repayment api

+++ database/migrations/2021_07_25_193952_create_loan_repayment_table.php
@@ -16,7 +16,7 @@ class CreateLoanRepaymentTable extends Migration
         Schema::create('loan_repayments', function (Blueprint $table) {
              $table->increments('id');
              $table->integer('loan_id')->unsigned();
-             $table->integer('amount')->unsigned();
+             $table->float('amount', 8, 2)->unsigned();
              $table->timestamps();
              
         });
